requirements.txt
Pillow==6.1.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
matplotlib==3.1.1